1-array_iterator.c

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -9,13 +9,13 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-int idx;
+	size_t idx;
 
-if (array != NULL && action != NULL)
-{
-for (idx = 0; idx < size; idx++)
-{
-action(array[idx]);
-}
-}
+	if (array != NULL && action != NULL)
+	{
+		for (idx = 0; idx < size; idx++)
+		{
+			action(array[idx]);
+		}
+	}
 }
